move common bitmask functions to a separate file
